Prep for interview

+++ ImageRename.Core/ImageRename.Core/ProcessFolder.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.IO;
 using System.Linq;
@@ -7,7 +6,6 @@ using ImageRename.Core.Model;
 
 namespace ImageRename.Core
 {
-
     public class ProcessFolder
     {
 
@@ -15,7 +13,7 @@ namespace ImageRename.Core
         public bool MoveToProcessedByYear { get; set; }
         public string ProcessedPath { get; set; }
 
-        private ObservableCollection<IImageFile> _images;
+        public ObservableCollection<IImageFile> Images;
         private string _rootFolder;
 
         #region RenameProgressEvent
@@ -47,10 +45,10 @@ namespace ImageRename.Core
         {
 
             var e = new ReportFindFilesProgressEventArgs();
-            if (_images != null)
+            if (Images != null)
             {
-                e.TotalFileCount = _images.Count();
-                e.FilesToRename = _images.Count(c => c.NeedsRenaming);
+                e.TotalFileCount = Images.Count();
+                e.FilesToRename = Images.Count(c => c.NeedsRenaming);
             }
 
             OnReportFilesFoundProgress(e);
@@ -63,8 +61,8 @@ namespace ImageRename.Core
                 throw new DirectoryNotFoundException(root);
             }
             _rootFolder = root;
-            _images = new ObservableCollection<IImageFile>();
-            _images.CollectionChanged += _images_CollectionChanged;
+            Images = new ObservableCollection<IImageFile>();
+            Images.CollectionChanged += _images_CollectionChanged;
             FindFiles(root);
             RenameFiles();
         }
@@ -76,23 +74,36 @@ namespace ImageRename.Core
 
         private void RenameFiles()
         {
-            if (_images == null || !_images.Any(a => a.NeedsRenaming == true))
+            if (Images == null || !Images.Any(a => a.NeedsRenaming == true))
             {
                 return;
             }
 
-            foreach (var item in _images.Where(w => w.NeedsRenaming == true))
+            foreach (var item in Images.Where(w => w.NeedsRenaming ))
             {
                 RenameFile(item);
+            }
+            foreach (var item in Images.Where(w=>w.NeedsMoving))
+            {
+                if(!File.Exists(item.FileDetails.FullName))
+                {
+                    continue;
+                }
+                File.Move(item.FileDetails.FullName, item.ProcessedFullName);
             }
         }
 
         private void RenameFile(IImageFile item)
         {
             var sourceFile = item.FileDetails.FullName;
-            var destinationFile = item.NewFilePath;
+            var destinationFile = item.ProcessedFullName;
             if (!DebugDontRenameFile)
             {
+                if (!File.Exists(item.FileDetails.FullName))
+                {
+                    throw new FileNotFoundException(item.FileDetails.FullName);
+                }
+                Helper.CreateDirectory(item.ProcessedDirectory, true);
                 File.Move(sourceFile, destinationFile);
             }
             item.FileDetails = new FileInfo(destinationFile);
@@ -104,9 +115,9 @@ namespace ImageRename.Core
 
         private void FindFiles(string root)
         {
-            if(MoveToProcessedByYear && !string.IsNullOrEmpty(ProcessedPath))
+            if (MoveToProcessedByYear && !string.IsNullOrEmpty(ProcessedPath))
             {
-                if(!Directory.Exists(ProcessedPath))
+                if (!Directory.Exists(ProcessedPath))
                 {
                     Directory.CreateDirectory(ProcessedPath);
                 }
@@ -123,12 +134,12 @@ namespace ImageRename.Core
                         case "jpg":
                         case "jpeg":
                         case "cr2":
-                            _images.Add(new ImageFile(file, ProcessedPath));
+                            Images.Add(new ImageFile(file, ProcessedPath));
                             break;
                         case "mov":
                         case "mp4":
                         case "m4a":
-                            _images.Add(new VideoFile(file, ProcessedPath));
+                            Images.Add(new VideoFile(file, ProcessedPath));
                             break;
                         default:
                             break;
